Refactor models out of db into models folder
